// src/shentuan/types/PaymentDetail.ts
export interface PaymentDetail {
    accountId: string;
    merchantId: string;
    merchantName: string;
    phoneNo: string;
    ownerName: string;
    accountName: string;
    accountNo: string;
    accountBankName: string;
    accountBankBranchName: string;
    status: number;
    remark: string;
    createTime: number;
    updateTime: number;
}